requirements.txt
streamlit>=1.28.0
arxiv>=2.1.0
sentence-transformers>=2.2.0,<2.3.0
# Using simple text matching instead of vector embeddings
google-generativeai>=0.3.0
ragas>=0.1.0
openai>=1.3.0
pypdf2>=3.0.0
requests>=2.31.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
langchain>=0.0.350
langchain-community>=0.0.2
pydantic>=2.5.0
datasets>=2.0.0
torch>=1.13.0,<2.1.0
transformers>=4.30.0,<4.40.0
